added optional fallback to suspense and transition as well
