+++ leptos/src/show.rs
@@ -1,7 +1,6 @@
-use leptos::{component, ChildrenFn};
-use leptos_dom::{IntoView, View};
+use leptos::{component, ChildrenFn, ViewFn};
+use leptos_dom::IntoView;
 use leptos_reactive::{create_memo, signal_prelude::*};
-use std::rc::Rc;
 
 /// A component that will show its children when the `when` condition is `true`,
 /// and show the fallback when it is `false`, without rerendering every time
@@ -39,7 +38,7 @@ pub fn Show<W>(
     when: W,
     /// A closure that returns what gets rendered if the when statement is false. By default this is the empty view.
     #[prop(optional, into)]
-    fallback: Fallback,
+    fallback: ViewFn,
 ) -> impl IntoView
 where
     W: Fn() -> bool + 'static,
@@ -49,38 +48,5 @@ where
     move || match memoized_when.get() {
         true => children().into_view(),
         false => fallback.run(),
-    }
-}
-
-/// New-type wrapper for the fallback view function to enable `#[prop(into, optional)]`.
-#[derive(Clone)]
-pub struct Fallback {
-    function: Rc<dyn Fn() -> View>,
-}
-
-impl Default for Fallback {
-    fn default() -> Self {
-        Self {
-            function: Rc::new(|| ().into_view()),
-        }
-    }
-}
-
-impl<F, IV> From<F> for Fallback
-where
-    F: Fn() -> IV + 'static,
-    IV: IntoView,
-{
-    fn from(value: F) -> Self {
-        Self {
-            function: Rc::new(move || value().into_view()),
-        }
-    }
-}
-
-impl Fallback {
-    /// Execute the wrapped function
-    pub fn run(&self) -> View {
-        (self.function)()
     }
 }
